Improve hover style
